Remove unecessary code
