+++ search_results.py
@@ -12,3 +12,30 @@ def create_search_results(book_items_list, search_results):
         # add volume to search results
         search_results[list_id] = volume_information
     #   under a key of id , offset from index to prevent a 0 list position
+
+
+def check_for_search_results(book_search_response):
+    results = None
+    try:
+        results = book_search_response["items"]
+    except KeyError:
+        results = None
+        print("I don't have any results for that search.")
+    return results
+
+
+def check_if_still_searching():
+    valid_response = ""
+
+    while valid_response == "":
+        valid_response = input(
+            """Would you like to try a different search? (Yes/No) \n""").lower()
+        if valid_response == "no":
+            break
+        elif valid_response == "yes":
+            break
+        else:
+            print(
+                "Looks like you entered a response other than Yes or No. Please try again.")
+            valid_response = ""
+    return valid_response
